ai_service/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
requests
python-multipart
psycopg2-binary
beautifulsoup4
apscheduler
pytz
google-generativeai